add empty test cases with test db setup
