+++ .gitignore
@@ -105,3 +105,5 @@ venv.bak/
 
 certs/
 migrations_local/
+
+tests/test.db
